finish

+++ include/GENE.h
@@ -14,11 +14,13 @@ class GENE {
         // 計算適值函數的值
         double calculate_adaptive_value();
         // 和誰交叉運算
+        // 單切點
         void cross_to1(GENE &who);
+        // 雙切點(默認)
         void cross_to2(GENE &who);
     // private:
         // 基因
-        vector<int> gene;
+        vector<double> gene;
         // 基因的適應值
         double adaptive_value;
         //
